app/requirements.txt
Django==1.7.1
Pillow==2.6.1
SQLAlchemy==0.9.8
dj-database-url==0.3.0
dj-static==0.0.6
django-model-utils==2.2
django-photologue==3.1.1
django-postgrespool==0.3.0
django-sortedm2m==0.8.1
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
wsgiref==0.1.2